refactor: consolidate version preview into TemplatePreviewHandler
Move previewVersion() from VersionComparisonHandler into the existing
TemplatePreviewHandler, keeping all preview logic in one place.
VersionComparisonHandler now only handles the compare dialog.
