+++ apps/epistola/src/main/kotlin/app/epistola/suite/handlers/TemplatePreviewHandler.kt
@@ -1,10 +1,12 @@
 package app.epistola.suite.templates
 
+import app.epistola.suite.documents.queries.PreviewDocument
 import app.epistola.suite.documents.queries.PreviewDraft
 import app.epistola.suite.generation.GenerationService
 import app.epistola.suite.htmx.templateId
 import app.epistola.suite.htmx.tenantId
 import app.epistola.suite.htmx.variantId
+import app.epistola.suite.htmx.versionId
 import app.epistola.suite.mediator.query
 import org.springframework.http.HttpHeaders
 import org.springframework.http.MediaType
@@ -97,6 +99,52 @@ class TemplatePreviewHandler(
                 )
         } catch (e: IllegalStateException) {
             // Template/variant/draft not found
+            return ServerResponse.notFound().build()
+        }
+
+        return ServerResponse.ok()
+            .contentType(MediaType.APPLICATION_PDF)
+            .header(HttpHeaders.CONTENT_DISPOSITION, "inline; filename=\"preview.pdf\"")
+            .body(pdfBytes)
+    }
+
+    /**
+     * Generates a PDF preview of a specific version (draft or published).
+     * Used by the version comparison dialog.
+     */
+    fun previewVersion(request: ServerRequest): ServerResponse {
+        val tenantId = request.tenantId()
+        val templateId = request.templateId(tenantId)
+            ?: return ServerResponse.badRequest().build()
+        val variantId = request.variantId(templateId)
+            ?: return ServerResponse.badRequest().build()
+        val versionId = request.versionId(variantId)
+            ?: return ServerResponse.badRequest().build()
+
+        val data: ObjectNode = try {
+            val body = request.body(String::class.java)
+            if (body.isBlank()) {
+                objectMapper.createObjectNode()
+            } else {
+                objectMapper.readTree(body) as ObjectNode
+            }
+        } catch (_: Exception) {
+            objectMapper.createObjectNode()
+        }
+
+        val pdfBytes = try {
+            PreviewDocument(
+                tenantId = tenantId.key,
+                templateId = templateId.key,
+                data = data,
+                variantId = variantId.key,
+                versionId = versionId.key,
+            ).query()
+        } catch (e: IllegalArgumentException) {
+            return ServerResponse.badRequest()
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(mapOf("error" to (e.message ?: "Validation failed")))
+        } catch (_: IllegalStateException) {
             return ServerResponse.notFound().build()
         }
